power: qpnp-charger: request fastchg falling irq
When device is in charging mode, if enable the revert boost regulator
for flash led, smbb will tranfer from fast_chg state to bst_on state,
the charging will stop. But usbin_valid && coarse_det are still valid,
and the battery_status not change to discharging.

Register a fastchg falling irq to update charging status and make the
charging icon consistent with charging status in this situation.

Change-Id: I0ce903f221d7550759b2601864c6a13a6036c494
Signed-off-by: Wu Fenglin <fenglinw@codeaurora.org>

Conflicts:
	drivers/power/qpnp-charger.c