power: qpnp-charger: add parallel ovp charge mode
This enables the parallel OVP charge mode for PM8941 devices.
This feature specific to PM8941 allows DC_IN and USB_IN to be
shorted on the device to allowing both input paths to be leveraged
and put more current into the battery.

Note that the feature is only enabled when fast charging
and disabled once fast charging is disabled.

Change-Id: Id76ef10c75eae267149437391f4e9fa6bfede18b
CRs-Fixed: 607941
Signed-off-by: David Keitel <dkeitel@codeaurora.org>

Conflicts:
	Documentation/devicetree/bindings/power/qpnp-charger.txt