misc: stm401: exit quickpeek workqueue when stm401 reset
In the event that quickwake detects a timeout the stm401 is reset
to attempt to recover. Included in the reset step is to flush the
quickpeek command list. If the quickpeek workqueue is running through
the list (and removing items) at the same time that the reset is
flushing the list a panic can occur when the same entry is deleted
from the list in both threads. To prevent this, the quickpeek
workqueue is prevented form running during an stm401 reset process.

Change-Id: I299a18903daa0b666b301484d22fa7244d7edbc2
Signed-off-by: Joe Swantek <jswantek@motorola.com>
Reviewed-on: http://gerrit.mot.com/660541
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Patrick Auchter <auchter@motorola.com>
Reviewed-by: Fred Fettinger <fettinge@motorola.com>
Reviewed-by: Jared Suttles <jsuttles@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
SLTApproved: Connie Zhao <czhao1@motorola.com>